self_d_python/requirements.txt
tensorflow==2.5.0
opencv-python==4.5.5.64
keyboard==0.13.5
mysql-connector-python==8.0.29
pandas==1.3.5
pyserial==3.5
protobuf==3.20.*